backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
langchain==0.1.9
langchain-openai==0.0.6
openai==1.12.0
# mysql-connector-python  # quando implementar o banco
pydantic==2.6.1
httpx==0.26.0
